# packages/contracts/openapi.yaml
openapi: 3.1.0
info:
  title: MergeCom API
  version: 0.3.0
  description: Identity, organization, RBAC, project, folder, and document metadata contracts.
paths:
...
  /v1/me:
    get:
      operationId: getCurrentUser
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Authenticated user and active organization context.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentUser' }
        '401': { $ref: '#/components/responses/ApiError' }
  /v1/session/organization:
    post:
      operationId: switchOrganization
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [organizationId]
              properties:
                organizationId: { type: string, format: uuid }
      responses:
        '200':
          description: Updated active organization context.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/CurrentUser' }
        '401': { $ref: '#/components/responses/ApiError' }
        '403': { $ref: '#/components/responses/ApiError' }
        '404': { $ref: '#/components/responses/ApiError' }
  /v1/organizations/{organizationId}/memberships:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
    get:
      operationId: listMemberships
      security: [{ sessionCookie: [] }]
      responses:
        '200':
          description: Memberships in the active organization.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [memberships]
                properties:
                  memberships:
                    type: array
                    items: { $ref: '#/components/schemas/Membership' }
        '401': { $ref: '#/components/responses/ApiError' }
        '403': { $ref: '#/components/responses/ApiError' }
        '404': { $ref: '#/components/responses/ApiError' }
  /v1/organizations/{organizationId}/invitations:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
    post:
      operationId: createInvitation
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [email, role]
              properties:
                email: { type: string, format: email, maxLength: 320 }
                role: { $ref: '#/components/schemas/OrganizationRole' }
                expiresInDays: { type: integer, minimum: 1, maximum: 14 }
                projectId: { type: string, format: uuid }
                projectRole: { $ref: '#/components/schemas/ProjectRole' }
      responses:
        '201':
          description: Invitation created and delivered.
          content:
            application/json:
              schema: { $ref: '#/components/schemas/Invitation' }
        '400': { $ref: '#/components/responses/ApiError' }
        '401': { $ref: '#/components/responses/ApiError' }
        '403': { $ref: '#/components/responses/ApiError' }
        '404': { $ref: '#/components/responses/ApiError' }
        '409': { $ref: '#/components/responses/ApiError' }
        '503': { $ref: '#/components/responses/ApiError' }
  /v1/invitations/accept:
    post:
      operationId: acceptInvitation
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [token]
              properties:
                token: { type: string, minLength: 20, maxLength: 256 }
      responses:
        '200':
          description: Invitation accepted once.
          content:
            application/json:
              schema:
                type: object
                additionalProperties: false
                required: [organizationId]
                properties:
                  organizationId: { type: string, format: uuid }
        '400': { $ref: '#/components/responses/ApiError' }
        '401': { $ref: '#/components/responses/ApiError' }
        '403': { $ref: '#/components/responses/ApiError' }
        '409': { $ref: '#/components/responses/ApiError' }
  /v1/organizations/{organizationId}/memberships/{membershipId}/role:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - $ref: '#/components/parameters/MembershipId'
    patch:
      operationId: changeMembershipRole
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [role]
              properties:
                role: { $ref: '#/components/schemas/OrganizationRole' }
      responses:
        '204': { description: Role changed. }
        '401': { $ref: '#/components/responses/ApiError' }
        '403': { $ref: '#/components/responses/ApiError' }
        '404': { $ref: '#/components/responses/ApiError' }
        '409': { $ref: '#/components/responses/ApiError' }
  /v1/organizations/{organizationId}/memberships/{membershipId}/status:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - $ref: '#/components/parameters/MembershipId'
    patch:
      operationId: changeMembershipStatus
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              additionalProperties: false
              required: [status]
              properties:
                status: { $ref: '#/components/schemas/MembershipStatus' }
      responses:
        '204': { description: Membership status changed. }
        '401': { $ref: '#/components/responses/ApiError' }
        '403': { $ref: '#/components/responses/ApiError' }
        '404': { $ref: '#/components/responses/ApiError' }
        '409': { $ref: '#/components/responses/ApiError' }
  /v1/organizations/{organizationId}/memberships/{membershipId}:
    parameters:
      - $ref: '#/components/parameters/OrganizationId'
      - $ref: '#/components/parameters/MembershipId'
    delete:
      operationId: removeMembership
      security: [{ sessionCookie: [] }]
      parameters:
        - $ref: '#/components/parameters/CsrfToken'
      responses:
        '204': { description: Membership removed. }
        '401': { $ref: '#/components/responses/ApiError' }
        '403': { $ref: '#/components/responses/ApiError' }
        '404': { $ref: '#/components/responses/ApiError' }
        '409': { $ref: '#/components/responses/ApiError' }
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: mergecom_session
  parameters:
    CsrfToken:
      name: X-CSRF-Token
      in: header
      required: true
      schema: { type: string }
    OrganizationId:
      name: organizationId
      in: path
      required: true
      schema: { type: string, format: uuid }
    MembershipId:
      name: membershipId
      in: path
      required: true
      schema: { type: string, format: uuid }
...
  responses:
    ApiError:
      description: Request rejected.
      content:
        application/json:
          schema: { $ref: '#/components/schemas/ApiError' }
  schemas:
    ApiError:
      type: object
      additionalProperties: false
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
    OrganizationRole:
      type: string
      enum:
        [
          owner,
          admin,
          project_lead,
          contributor,
          reviewer,
          viewer,
          external_reviewer,
        ]
    MembershipStatus:
      type: string
      enum: [active, suspended]
    ProjectRole:
      type: string
      enum: [project_lead, contributor, reviewer, viewer]
...
    Organization:
      type: object
      additionalProperties: false
      required: [id, name, role]
      properties:
        id: { type: string, format: uuid }
        name: { type: string }
        role: { $ref: '#/components/schemas/OrganizationRole' }
    CurrentUser:
      type: object
      additionalProperties: false
      required: [user, activeOrganization, organizations, session]
      properties:
        user:
          type: object
          additionalProperties: false
          required: [id, displayName, email, emailVerified]
          properties:
            id: { type: string, format: uuid }
            displayName: { type: string }
            email: { type: string, format: email }
            emailVerified: { type: boolean }
        activeOrganization:
          oneOf:
            - type: 'null'
            - type: object
              additionalProperties: false
              required: [id, name, role, status]
              properties:
                id: { type: string, format: uuid }
                name: { type: string }
                role: { $ref: '#/components/schemas/OrganizationRole' }
                status: { $ref: '#/components/schemas/MembershipStatus' }
        organizations:
          type: array
          items: { $ref: '#/components/schemas/Organization' }
        session:
          type: object
          additionalProperties: false
          required: [csrfToken, expiresAt]
          properties:
            csrfToken: { type: string }
            expiresAt: { type: string, format: date-time }
    Membership:
      type: object
      additionalProperties: false
      required: [id, userId, name, email, role, status, joinedAt]
      properties:
        id: { type: string, format: uuid }
        userId: { type: string, format: uuid }
        name: { type: string }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/OrganizationRole' }
        status: { $ref: '#/components/schemas/MembershipStatus' }
        joinedAt: { type: string, format: date-time }
    Invitation:
      type: object
      additionalProperties: false
      required: [id, email, role, expiresAt]
      properties:
        id: { type: string, format: uuid }
        email: { type: string, format: email }
        role: { $ref: '#/components/schemas/OrganizationRole' }
        expiresAt: { type: string, format: date-time }
        acceptanceUrl: { type: string, format: uri }
        projectId: { type: string, format: uuid }
        projectRole: { $ref: '#/components/schemas/ProjectRole' }
...